Attempt to simplify different types of handlers
I think the numbering implied that the different methods were mutually exclusive, even though there was text that said that they were not.

I don't think the explicit `send-type` example is necessary, it seems like a more awkward alternative to just returning the data.  Folks who really want that for some reason can figure it out by following the `send-type` link and combining it with the native API example.